fix: move DevToolData & RegisterBuildPlatformResult to modern/plugin
