+++ packages/solutions/app-tools/src/types/hooks.ts
@@ -3,6 +3,8 @@ import type {
   AsyncWaterfall,
   AsyncWorkflow,
   ParallelWorkflow,
+  RegisterBuildPlatformResult,
+  DevToolData,
 } from '@modern-js/plugin';
 import type {
   Entrypoint,
@@ -13,7 +15,6 @@ import type {
   ServerRoute,
 } from '@modern-js/types';
 import type { Stats, MultiStats } from '@modern-js/builder-shared';
-import { DevOptions } from '../utils/types';
 
 export interface ImportSpecifier {
   local?: string;
@@ -102,25 +103,3 @@ export type AppToolsHooks = {
   registerBuildPlatform: ParallelWorkflow<void, RegisterBuildPlatformResult>;
   beforeBuildPlatform: ParallelWorkflow<RegisterBuildPlatformResult[], void>;
 };
-
-export interface RegisterBuildPlatformResult {
-  platform: string | string[];
-  build: (
-    currentPlatform: string,
-    context: { isTsProject: boolean },
-  ) => void | Promise<void>;
-}
-
-export interface DevToolData {
-  name: string;
-  subCommands?: string[];
-  menuItem?: {
-    name: string;
-    value: string;
-  };
-  // disableRunBuild?: boolean;
-  action: (
-    options: DevOptions,
-    context: { isTsProject?: boolean },
-  ) => void | Promise<void>;
-}
