+++ packages/solutions/module-tools/src/hooks/build.ts
@@ -1,10 +1,13 @@
-import { createParallelWorkflow, createAsyncPipeline } from '@modern-js/plugin';
+import {
+  createParallelWorkflow,
+  createAsyncPipeline,
+  RegisterBuildPlatformResult,
+} from '@modern-js/plugin';
 import { BuildCommandOptions } from '../types';
 import type { BuildConfig, BaseBuildConfig } from '../types/config';
 import type {
   BuildTaskResult,
   BuildResult,
-  RegisterBuildPlatformResult,
   BuildPlatformResult,
 } from '../types/hooks';
