+++ packages/solutions/module-tools/src/types/hooks.ts
@@ -1,3 +1,4 @@
+import { DevToolData as BaseDevToolData } from '@modern-js/plugin';
 import { registerHook } from '../hooks';
 import type { BaseBuildConfig, BuildConfig } from './config';
 import type { DevCommandOptions } from './command';
@@ -13,31 +14,12 @@ export interface BuildResult {
   message?: string;
   config: BuildConfig;
 }
-export interface RegisterBuildPlatformResult {
-  platform: string | string[];
-  build: (
-    currentPlatform: string,
-    context: { isTsProject: boolean },
-  ) => void | Promise<void>;
-}
 export interface BuildPlatformResult {
   status: 'success' | 'fail';
   message: string;
 }
 
-export interface DevToolData {
-  name: string;
-  subCommands?: string[];
-  menuItem?: {
-    name: string;
-    value: string;
-  };
-  disableRunBuild?: boolean;
-  action: (
-    options: DevCommandOptions,
-    context: { isTsProject?: boolean },
-  ) => void | Promise<void>;
-}
+export type DevToolData = BaseDevToolData<DevCommandOptions>;
 
 export type PromptResult = { choiceDevTool: string | symbol } & Record<
   string,
