+++ packages/toolkit/plugin/src/manager/types.ts
@@ -102,3 +102,25 @@ export type PluginOptions<
 export type CommonAPI<Hooks> = {
   useHookRunners: () => ToRunners<Hooks>;
 };
+
+export interface DevToolData<DevOptions = any> {
+  name: string;
+  subCommands?: string[];
+  menuItem?: {
+    name: string;
+    value: string;
+  };
+  disableRunBuild?: boolean;
+  action: (
+    options: DevOptions,
+    context: { isTsProject?: boolean },
+  ) => void | Promise<void>;
+}
+
+export interface RegisterBuildPlatformResult {
+  platform: string | string[];
+  build: (
+    currentPlatform: string,
+    context: { isTsProject: boolean },
+  ) => void | Promise<void>;
+}
